Add randomized pick-any question and fix pick_randomly (#51)
Add randomized and correct-count options to the pick-any directive.
They are used to define checkbox questions that randomly select
N options to display at a time (N = the randomized value).
The correct-count option defines how many selected options are
correct.

Question titles may use the mark {#} instead of a number.
MOOC-grader numbers the question then. This is intended for
pick-randomly questionnaires so that the MOOC-grader may number
the questions correctly after randomly selecting them.

Randomized questionnaires no longer include the data-aplus-quiz
attribute in the HTML. This fixes the issue that the questionnaire
could not be submitted again because the feedback form includes
a used, invalid nonce value. A+ would only show the feedback
of the questionnaire instead of the exercise description when
the data-aplus-quiz attribute is used.

Co-authored-by: Markku Riekkinen <markku.riekkinen@aalto.fi>